Move error overlay into generalized DecryptFragment, better error handling in DecryptTextActivity
